Only starting the server when the user requests it
Starting it by default every time the user opens a markdown file is... unwise.
This way the user has a command he can call when he wants the server to be
started, and if there's a desire for the previous, automatic behavior can added
to one's vimrc like so:

  autocmd BufEnter *.{md,mkd,mkdn,mark*} :StartMarkdownServer

So this is left up to the user.

Also, some minor tweaks were added to the code; function prefixes were replaced
with 's:', which makes the functions script-local, thus obviating the need for a
function prefix on globally accessible functions.